assign teacher to courses feature complete

+++ lib/features/admin/shared/courses/data/models/SemesterCourseModel.dart
@@ -32,7 +32,7 @@ class SemesterCourseModel extends SemesterCourse{
       courseDept: map['courseDept'] as String,
       courseType: map['courseType'] as String,
       courseSemester: map['courseSemester'] as String,
-      courseCreditHours: map['courseCreditHours'].toDouble(),
+      courseCreditHours: map['courseCreditHours'].toInt(),
     );
   }
